pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.66.0",
    "pillow>=11.1.0",
    "playwright>=1.50.0",
    "streamlit>=1.43.1",
]
